requirements.txt
python-dotenv
streamlit

langchain
langchain-ollama
